Changes to include images, renamed first three projects
Changes have been made to about.html, contact.html, project.html, and index.html so that images are pulled from the /public/images directory. Three of the existing projects were renamed and images were added for powerwheels, ping pong shooter, and the rover.

TODO: resize the images and add them back in. These are too big to actually include in the portfolio.